% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_vago_fonts.R
\name{install_vago_fonts}
\alias{install_vago_fonts}
\title{Install VAGO Fonts}
\usage{
install_vago_fonts(font.path = "C:/windows/fonts")
}
\description{
Themes in this package use the Segoe UI font.
}
\details{
This function allows these to be installed within the package
}
\seealso{
Other vagothemes: 
\code{\link{theme_vago_map}()}
}
\concept{vagothemes}
