% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_vago_lilac.R
\name{theme_vago_lilac}
\alias{theme_vago_lilac}
\title{A ggplot2 theme with background colour lilac and text colours to black.}
\usage{
theme_vago_lilac(
  base_colour = vago.lilac,
  base_size = 15,
  text_colour = "black"
)
}
\description{
Colours are sourced from \code{vagotheme} and are vago style guide compliant.
}
\seealso{
Other vagotheme: 
\code{\link{theme_vago_foundation}()},
\code{\link{theme_vago_white}()}
}
\concept{vagotheme}
