% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_vago_map.R
\name{theme_vago_map}
\alias{theme_vago_map}
\title{VAGO Map Theme}
\usage{
theme_vago_map(
  base_colour = c("white", "lilac"),
  base_size = 15,
  base_family = "Noto Sans"
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
This theme is designed to be be used when creating maps.
}
\details{
This theme takes \code{\link[ggthemes]{theme_map}()} and sets
base text family and size to \verb{Noto Sans} and \code{15}.
}
\seealso{
Other vagothemes: 
\code{\link{install_vago_fonts}()}
}
\concept{vagothemes}
