% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_vago_white.R
\name{theme_vago_white}
\alias{theme_vago_white}
\title{A ggplot2 theme with background colour white and text colours to black.}
\usage{
theme_vago_white(base_size = 15, text_colour = "black")
}
\arguments{
\item{base_size}{a numeric with default 15}

\item{text_colour}{a (character) string with default "Black"}
}
\description{
Colours are sourced from \code{vagotheme} and are vago style guide compliant.
}
\seealso{
Other vagotheme: 
\code{\link{theme_vago_foundation}()},
\code{\link{theme_vago_lilac}()}
}
\concept{vagotheme}
