% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.blue}
\alias{vago.blue}
\title{VAGO Blue}
\format{
\subsection{\code{vago.blue}}{

An object of class \code{colour}.
}
}
\source{
\url{https://github.com/zerogetsamgow/vagotheme}
}
\usage{
vago.blue
}
\description{
Blue is used in majority
}
\examples{
plot(vago.blue)
}
\seealso{
\code{\link[=pal_colour]{pal_colour()}}
}
\keyword{datasets}
