% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.cerulean}
\alias{vago.cerulean}
\title{VAGO Cerulean}
\format{
\subsection{\code{vago.cerulean}}{

An object of class \code{colour} with hex value \verb{#2F8DF8}.
}
}
\source{
\url{https://github.com/zerogetsamgow/vagotheme}
}
\usage{
vago.cerulean
}
\description{
VAGO Cerulean is a base and theme colour.
}
\examples{
barplot(c(2,5,3), col=rep(vago.cerulean,3))
}
\keyword{datasets}
