% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.lightteal}
\alias{vago.lightteal}
\title{VAGO Light Teal}
\format{
\subsection{\code{vago.lightteal}}{

An object of class \code{colour} with hex value \verb{#A3E9E9}.
}
}
\source{
\url{https://github.com/zerogetsamgow/vagotheme}
}
\usage{
vago.lightteal
}
\description{
VAGO Light Teal is a lighter shade of the Secondary colour
VAGO Light teal is a base, theme and chart colour.
}
\examples{
barplot(c(2,5,3), col=rep(vago.lightteal,3))
}
\keyword{datasets}
