% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.lilac}
\alias{vago.lilac}
\title{VAGO Lilac}
\format{
\subsection{\code{vago.lilac}}{

An object of class \code{colour} with hex value \verb{#E3E3ED}.
}
}
\usage{
vago.lilac
}
\description{
VAGO Lilac is a background colour.
}
\examples{
plot(vago.lilac)
}
\keyword{datasets}
