% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.purple}
\alias{vago.purple}
\title{VAGO Purple}
\format{
\subsection{\code{vago.purple}}{

An object of class \code{colour} with hex value \verb{#2D1A78}.
}
}
\source{
\url{https://github.com/zerogetsamgow/vagotheme}
}
\usage{
vago.purple
}
\description{
VAGO Purle is a base and theme colour.
}
\examples{
barplot(c(2,5,3), col=rep(vago.purple,3))
}
\keyword{datasets}
