% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.red}
\alias{vago.red}
\title{VAGO Red}
\format{
\subsection{\code{vago.red}}{

An object of class \code{colour} with hex value \verb{#D60000}.
}
}
\usage{
vago.red
}
\description{
VAGO Red is a base, theme and traffic colour.
}
\examples{
barplot(c(2,5,3), col=rep(vago.red,3))
}
\keyword{datasets}
