% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.teal}
\alias{vago.teal}
\title{vago Teal}
\format{
\subsection{\code{vago.teal}}{

An object of class \code{colour}.
}
}
\usage{
vago.teal
}
\description{
Teal is the secondary colour.
}
\examples{
plot(vago.teal)
}
\keyword{datasets}
