% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.white}
\alias{vago.white}
\title{VAGO White}
\format{
\subsection{\code{vago.white}}{

An object of class \code{colour} with hex value \verb{#AAAAAA}.
}
}
\source{
\url{https://github.com/zerogetsamgow/vagotheme}
}
\usage{
vago.white
}
\description{
VAGO white is a background colour.
}
\examples{
barplot(c(2,5,3), col=rep(vago.white,3))
}
\keyword{datasets}
