% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.amber}
\alias{vago.amber}
\title{VAGO Amber}
\format{
\subsection{\code{vago.amber}}{

An object of class \code{colour} with hex value \verb{#FFC000}.
}
}
\source{
\url{https://github.com/zerogetsamgow/vagotheme}
}
\usage{
vago.amber
}
\description{
VAGO Amber is a traffic colour.
}
\examples{
barplot(c(2,5,3), col=rep(vago.amber,3))
}
\keyword{datasets}
