% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.blue}
\alias{vago.blue}
\title{VAGO Blue}
\format{
\subsection{\code{vago.blue}}{

An object of class \code{colour} with hex value \verb{#054894}.
}
}
\source{
\url{https://github.com/zerogetsamgow/vagotheme}
}
\usage{
vago.blue
}
\description{
VAGO Blue is the primary colour.
VAGO Blue is a base and theme colour.
}
\examples{
barplot(c(2,5,3), col=rep(vago.blue,3))
}
\seealso{
\code{\link[=pal_colour]{pal_colour()}}
}
\keyword{datasets}
