% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.darkblue}
\alias{vago.darkblue}
\title{VAGO Dark Blue}
\format{
\subsection{\code{vago.darkblue}}{

An object of class \code{colour} with hex value \verb{#04366F}.
}
}
\source{
\url{https://github.com/zerogetsamgow/vagotheme}
}
\usage{
vago.darkblue
}
\description{
VAGO Dark blue is a darker shade of the primary colour
VAGO Dark blue is a chart colour.
}
\examples{
plot(vago.darkblue)
}
\seealso{
\code{\link[=pal_colour]{pal_colour()}}
}
\keyword{datasets}
