% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.darkteal}
\alias{vago.darkteal}
\title{VAGO Dark Teal}
\format{
\subsection{\code{vago.darkteal}}{

An object of class \code{colour} with hex value \verb{#208888}.
}
}
\source{
\url{https://github.com/zerogetsamgow/vagotheme}
}
\usage{
vago.darkteal
}
\description{
VAGO Dark Teal is a darker shade of the Secondary colour
VAGO Dark Teal is a base, theme and chart colour.
}
\examples{
barplot(c(2,5,3), col=rep(vago.darkteal,3))
}
\keyword{datasets}
