% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.green}
\alias{vago.green}
\title{VAGO green}
\format{
\subsection{\code{vago.green}}{

An object of class \code{colour} with hex value \verb{#00A86B}.
}
}
\usage{
vago.green
}
\description{
VAGO Green is a base, theme and traffic colour.
}
\examples{
barplot(c(2,5,3), col=rep(vago.green,3))
}
\keyword{datasets}
