% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.grey}
\alias{vago.grey}
\title{VAGO Grey}
\format{
\subsection{\code{vago.grey}}{

An object of class \code{colour} with hex value \verb{#BFBFBF}.
}
}
\usage{
vago.grey
}
\description{
VAGO Lilac is a chart colour.
}
\examples{
plot(vago.grey)
}
\keyword{datasets}
