% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.magenta}
\alias{vago.magenta}
\title{VAGO Magenta}
\format{
\subsection{\code{vago.magenta}}{

An object of class \code{colour} with hex value \verb{#930073}.
}
}
\usage{
vago.magenta
}
\description{
VAGO Magenta is a base and theme colour.
}
\examples{
plot(vago.magenta)
}
\keyword{datasets}
