% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vago_colours.R
\docType{data}
\name{vago.teal}
\alias{vago.teal}
\title{VAGO Teal}
\format{
\subsection{\code{vago.teal}}{

An object of class \code{colour} with hex value \verb{#2BB5B6}.
}
}
\source{
\url{https://github.com/zerogetsamgow/vagotheme}
}
\usage{
vago.teal
}
\description{
Teal is the secondary colour.
VAGO teal is a base, theme and chart colour.
}
\examples{
barplot(c(2,5,3), col=rep(vago.teal,3))
}
\keyword{datasets}
